+++ client/config.js
@@ -1,9 +1,9 @@
 module.exports = {
-  WORLD_MAX_X: 1000,
+  WORLD_MAX_X: 200,
   WORLD_MAX_Y: 100,
   WINDOW_DEFAULT_WIDTH: 100,
   WINDOW_DEFAULT_HEIGHT: 100,
-  VIEWPORT_COUNT: 10,
+  VIEWPORT_COUNT: 2,
   FISH_COUNT: 20,
   FISH_RESTING_SPEED: 5,
   FISH_STARTLE_SPEED: 50,
