+++ client/index.js
@@ -1,6 +1,6 @@
 var viewport = require('./viewport');
-var school = require('../shared/school');
-var sub = require('../shared/sub');
+var school = require('./school');
+var sub = require('./sub');
 var messages = require('./messages');
 var schoolSync = require('./school-sync');
 var animate = require('./animate');
