+++ client/random.js
